lib/main: add reactor_{run,stop} and use them in main
Problem: the main binary hangs indefinitely waiting for the reactor to
stop, and there are no safe interfaces to reactor_{run,stop}.

Solution: add safe interfaces for reactor run/stop and then use them in
the main function.  Also drop the `future.wait_for` since we are going
the async reactor route.

Note: unfortunately, it appears that not entering the reactor and just
going the declarative route does not work :(